requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=9.0.0